Fix: dotnet pack
This project has an fairly lengthy build script with a lot of options, but 'dotnet build' and 'dotnet test' work fine. 'dotnet pack' throws errors because of the format of the csproj file (unsupported license reference, double add of icon). This fixes that and puts 0.0.12 in the package version too.